requirements.txt
streamlit==1.35.0
tensorflow==2.16.1
pandas==2.2.2
fpdf2==2.7.8
opencv-python-headless==4.9.0.80
Pillow==10.3.0